refactor: actually its probably better if they stay as server/client
