+++ src/lib/communication/mod.rs
@@ -1,12 +1,12 @@
 //! Common infrastructure to pass messages between `swd` and `swctl`.
 
-pub mod request;
+pub mod client;
 pub mod delete;
 pub mod details;
 pub mod info;
 pub mod lap;
 pub mod pause;
 pub mod play;
-pub mod reply;
+pub mod server;
 pub mod start;
 pub mod stop;
